moved config to config module
