+++ config/config.py
@@ -1,3 +1,54 @@
-from config.sah_config import SAHConfig, get_db_credentials_path
+# MIT License
+#
+# Copyright (c) 2020-2024 Send A Hug
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# The provided Software is separate from the idea behind its website. The Send A Hug
+# website and its underlying design and ideas are owned by Send A Hug group and
+# may not be sold, sub-licensed or distributed in any way. The Software itself may
+# be adapted for any purpose and used freely under the given conditions.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+import os
+from pathlib import Path
 
-sah_config = SAHConfig(credentials_path=get_db_credentials_path())
+from config.sah_config import SAHConfig
+
+HOME = Path(os.environ.get("HOME", ""))
+SAH_HOME = Path(os.environ.get("SAH_HOME", HOME / "git" / "send-hug-backend"))
+SECRETS_PATH = SAH_HOME / ".secrets"
+FIREBASE_CREDENTIALS_FILE = Path(
+    os.environ.get(
+        "FIREBASE_CREDENTIALS_FILE",
+        SECRETS_PATH / "platform_firebase_credentials" / "latest.json",
+    )
+)
+
+
+def get_db_credentials_path() -> Path:
+    return Path(
+        os.environ.get(
+            "DB_CREDENTIALS_PATH", SECRETS_PATH / "db_development_creds" / "latest.json"
+        )
+    )
+
+
+sah_config = SAHConfig(
+    credentials_path=get_db_credentials_path(),
+    certificate_path=FIREBASE_CREDENTIALS_FILE,
+)
